backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
google-cloud-documentai==2.32.0
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
python-multipart==0.0.20
httpx==0.28.1
neo4j==5.27.0
rapidfuzz==3.10.1
tavily-python==0.5.0
pytest==8.3.4
pytest-asyncio==0.24.0
